refactor: ProductResponse 패키지 이동
